+++ models/item_model.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Float, Integer, String, func
+from sqlalchemy import Column, Integer, Numeric, String, func
 from sqlalchemy.dialects.postgresql import TIMESTAMP
 from sqlalchemy_serializer import SerializerMixin
 
@@ -14,7 +14,7 @@ class ItemModel(Base, SerializerMixin):
     product = Column(String, nullable=False)
     name = Column(String, nullable=False)
     description = Column(String, nullable=False)
-    price = Column(Float, nullable=False)
+    price = Column(Numeric(10, 2), nullable=False)
 
     created_at = Column(TIMESTAMP(timezone=True), server_default=func.now())
     updated_at = Column(TIMESTAMP(timezone=True), server_default=func.now(), onupdate=func.now())
